+++ console/src/components/Common/utils/pgUtils.tsx
@@ -27,10 +27,12 @@ export interface BaseTableColumn {
 }
 
 export interface TableColumn extends BaseTableColumn {
-  column_name: string;
-  data_type: string;
   udt_name: string;
   column_default: string;
+  is_generated?: string;
+  is_nullable?: string;
+  is_identity?: string;
+  identity_generation?: 'ALWAYS' | 'BY DEFAULT' | null;
 }
 
 export type ForeignKeyConstraint = {